backend/requirements.txt
flask==2.3.3
openai-whisper==20231214
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3
yt-dlp==2024.01.01
pydub==0.25.1

